Create Search_oM and Search_Engine to prepare for migration to BHoM_Engine (can easily be renamed if needed)
