+++ AGS_Toolkit.sln
@@ -1,13 +1,16 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32901.82
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AGS_Adapter", "AGS_Adapter\AGS_Adapter.csproj", "{F84BEFC2-C702-4A86-9626-C7FA48785606}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AGS_Engine", "AGS_Engine\AGS_Engine.csproj", "{C68B4D77-E46D-41C9-A72C-F29984D8AB4A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AGS_Engine", "AGS_Engine\AGS_Engine.csproj", "{C68B4D77-E46D-41C9-A72C-F29984D8AB4A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AGS_oM", "AGS_oM\AGS_oM.csproj", "{FA23407E-1C08-4183-A02E-A57FA9D09BCA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AGS_oM", "AGS_oM\AGS_oM.csproj", "{FA23407E-1C08-4183-A02E-A57FA9D09BCA}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Search_oM", "Search_oM\Search_oM.csproj", "{E1CEB00E-A4A7-4294-A8E2-54E45B1A84D3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Search_Engine", "Search_Engine\Search_Engine.csproj", "{17193F05-E8EB-450D-A0E7-5B5B7FB09B55}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +30,14 @@ Global
 		{FA23407E-1C08-4183-A02E-A57FA9D09BCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FA23407E-1C08-4183-A02E-A57FA9D09BCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FA23407E-1C08-4183-A02E-A57FA9D09BCA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E1CEB00E-A4A7-4294-A8E2-54E45B1A84D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E1CEB00E-A4A7-4294-A8E2-54E45B1A84D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E1CEB00E-A4A7-4294-A8E2-54E45B1A84D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E1CEB00E-A4A7-4294-A8E2-54E45B1A84D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{17193F05-E8EB-450D-A0E7-5B5B7FB09B55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{17193F05-E8EB-450D-A0E7-5B5B7FB09B55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{17193F05-E8EB-450D-A0E7-5B5B7FB09B55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{17193F05-E8EB-450D-A0E7-5B5B7FB09B55}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
