+++ Search_Engine/Ratios/TokenAbbreviationRatio.cs
@@ -21,19 +21,15 @@
  */
 
 using BH.oM.Base;
-using BH.oM.Adapters.AGS;
 using BH.oM.Base.Attributes;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using FuzzySharp;
-using FuzzySharp.Extractor;
-using FuzzySharp.SimilarityRatio;
-using FuzzySharp.SimilarityRatio.Scorer;
-using FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+using FuzzySharp.PreProcess;
 
-namespace BH.Engine.Adapters.AGS
+namespace BH.Engine.Search
 {
     public static partial class Compute
     {
@@ -41,40 +37,14 @@ namespace BH.Engine.Adapters.AGS
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Carries out a fuzzy match of the two strings provided using the scorer specified.")]
+        [Description("Similarity ratio that attempts to determine whether one strings tokens are an abbreviation of the other strings tokens. One string must have all its " +
+            "characters in order in the other string to even be considered. This makes use of the FuzzySharp library.")]
         [Input("text", "The string to carry out the fuzzy matching on.")]
         [Input("compare", "The string to compare against.")]
-        [Input("scorer", "The method to use to score the strings when compared.")]
         [Output("r", "The ratio of similarity between the two strings.")]
-        public static int FuzzyMatching(string text, string compare, Scorer scorer)
+        public static int TokenAbbreviationRatio(string text, string compare)
         {
-            switch (scorer)
-            {
-                case Scorer.DefaultRatioScorer:
-                default:
-                    return SimpleRatio(text, compare);
-                case Scorer.PartialRatioScorer:
-                    return PartialRatio(text, compare);
-                case Scorer.TokenSetScorer:
-                    return TokenSetRatio(text, compare);
-                case Scorer.PartialTokenSetScorer:
-                    return PartialTokenSetRatio(text, compare);
-                case Scorer.TokenSortScorer:
-                    return TokenSortRatio(text, compare);
-                case Scorer.PartialTokenSortScorer:
-                    return PartialTokenSortRatio(text, compare);
-                case Scorer.TokenAbbreviationScorer:
-                    return TokenAbbreviationRatio(text, compare);
-                case Scorer.PartialTokenAbbreviationScorer:
-                    return PartialTokenAbbreviationRatio(text, compare);
-                case Scorer.TokenInitialismScorer:
-                    return TokenInitialismRatio(text, compare);
-                case Scorer.PartialTokenInitialismScorer:
-                    return PartialTokenInitialismRatio(text, compare);
-                case Scorer.WeightedRatioScorer:
-                    return WeightedRatio(text, compare);
-
-            }
+            return Fuzz.TokenAbbreviationRatio(text, compare, PreprocessMode.Full);
         }
 
         /***************************************************/
