+++ Search_Engine/Ratios/TokenSetRatio.cs
@@ -28,7 +28,7 @@ using System.ComponentModel;
 using System.Linq;
 using FuzzySharp;
 
-namespace BH.Engine.Adapters.AGS
+namespace BH.Engine.Search
 {
     public static partial class Compute
     {
@@ -36,13 +36,14 @@ namespace BH.Engine.Adapters.AGS
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Calculates a weighted ratio between the different algorithms for best results. This makes use of the FuzzySharp library.")]
+        [Description("Splits the strings into tokens and computes intersections and remainders between the tokens of the two strings.A comparison string is then " +
+            "built up and is compared using the simple ratio algorithm. Useful for strings where words appear redundantly. This makes use of the FuzzySharp library.")]
         [Input("text", "The string to carry out the fuzzy matching on.")]
         [Input("compare", "The string to compare against.")]
         [Output("r", "The ratio of similarity between the two strings.")]
-        public static int WeightedRatio(string text, string compare)
+        public static int TokenSetRatio(string text, string compare)
         {
-            return Fuzz.WeightedRatio(text, compare);
+            return Fuzz.TokenSetRatio(text, compare);
         }
 
         /***************************************************/
