+++ Search_Engine/Ratios/WeightedRatio.cs
@@ -27,10 +27,8 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using FuzzySharp;
-using FuzzySharp.PreProcess;
 
-
-namespace BH.Engine.Adapters.AGS
+namespace BH.Engine.Search
 {
     public static partial class Compute
     {
@@ -38,14 +36,13 @@ namespace BH.Engine.Adapters.AGS
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Similarity ratio that attempts to determine whether one strings tokens are an abbreviation of the other strings tokens. One string must have all its " +
-            "characters in order in the other string to even be considered. This makes use of the FuzzySharp library.")]
+        [Description("Calculates a weighted ratio between the different algorithms for best results. This makes use of the FuzzySharp library.")]
         [Input("text", "The string to carry out the fuzzy matching on.")]
         [Input("compare", "The string to compare against.")]
         [Output("r", "The ratio of similarity between the two strings.")]
-        public static int PartialTokenAbbreviationRatio(string text, string compare)
+        public static int WeightedRatio(string text, string compare)
         {
-            return Fuzz.PartialTokenAbbreviationRatio(text, compare, PreprocessMode.Full);
+            return Fuzz.WeightedRatio(text, compare);
         }
 
         /***************************************************/
